honey/ssh_honeypy/requirements.txt
dash==2.17.1
dash_bootstrap_components==1.6.0
dash_bootstrap_templates==1.2.3
Flask==3.0.3
pandas==2.2.2
paramiko==3.4.0
plotly==5.23.0
python-dotenv==1.0.1
Requests==2.32.3
